+++ src/middleware/util.h
@@ -1,6 +1,6 @@
 /*
 author          Oliver Blaser
-date            10.01.2023
+date            26.02.2023
 copyright       GNU GPLv3 - Copyright (c) 2023 Oliver Blaser
 */
 
@@ -13,6 +13,28 @@ copyright       GNU GPLv3 - Copyright (c) 2023 Oliver Blaser
 
 namespace util
 {
+    class FileCounter
+    {
+    public:
+        using counter_type = size_t;
+
+    public:
+        FileCounter() : m_total(0), m_copied(0) {}
+        virtual ~FileCounter() {}
+
+        FileCounter& add(counter_type total, counter_type copied) { m_total += total; m_copied += copied; return (*this); }
+        FileCounter& add(const FileCounter& other) { return add(other.total(), other.copied()); }
+        FileCounter& addTotal(counter_type value = 1) { m_total += value; return (*this); }
+        FileCounter& addCopied(counter_type value = 1) { m_copied += value; return (*this); }
+
+        const counter_type& total() const { return m_total; }
+        const counter_type& copied() const { return m_copied; }
+
+    private:
+        counter_type m_total;
+        counter_type m_copied;
+    };
+
     class ResultCounter
     {
     public: