+++ release-notes.md
@@ -4,10 +4,11 @@
 
 ### v1.0.1
 
-- Added VID_* files in Huawai scheme (commit 83e5515bcee712595b53a225e7339f2e488d1cb0)
+- Added VID_* files in Huawai scheme (commit [83e5515](83e5515bcee712595b53a225e7339f2e488d1cb0))
+- Added file counters
 
 
 
 ### v1.0.0
 
-initial release
+Initial release